Refine Bottom Sheets in/out transition (#14831)
As specified in https://github.com/wordpress-mobile/gutenberg-mobile/issues/595
I did my best to adjust to the specs there, and for the most part the animations
matches what was requested.

One limitation is that due to how react-native-modal is implemented, we don't
know the view's height when it's about to be presented, and so it starts the
"slide in" animation not from the bottom edge, but from a lower point, since it
uses the window height to displace the view and not the view's height. The
result is an animation that appears slower than specified, since the view has
to traverse a longer distance in the same time, and there is a delay before the
view starts appearing.

This is mildly annoying, but I could not find a way to make it work without
rewriting react-native-modal.