Fix @wordpress/data documentation (#15831)
* Make import more conservative in sample code

As `data` is only used on the next line in the example, it seems preferable to not reserve that variable name for the scope of the entire file.

* Fix syntax errors and improper object handling in sample code

The `existingActions`/`existingSelectors` objects were being treated like arrays with `.map()`. This commit changes the code to treat them like plain JS objects.