Selecting Parent Blocks: Try clickthrough (#15537)
* Selecting parents: Try clickthrough.

Clickthrough has you select the parent before you can select the child. This is already in place on the mobile breakpoints, this just expands it to desktop as well.

It is a work in progress, right now it is not working as intended: once you have "unlocked" the deepest level, it becomes immediately locked and you have to click through the layers again to unlock it again. The deepest layer should always be unlocked until you deselect all blocks again.

* Render overlay on top of inner block only when none of the nested blocks is selected

* Fix overlay, fix breadcrumb, polish.

* Remove click-overlay.

* Fix the selection of inner blocks for reusable blocks template

* Disable async mode for parent blocks when nested block selected

* Refactor BlockListBlock to use AsyncModeProvider

* Make the reusable blocks save button clickable. i

At least this means you can edit and save reusable blocks.

* Fix so reusable blocks with nesting are editable.

* Fix movers.

The z-index was too low. It had to be higher to accommodate the other z-index changes.

* Bring the behavior closer to what we have as of today