Fix columns full-wide regression. (#18021)
The Columns block, when full-wide, has intentional left and right padding to ensure the mover controls of child blocks are accessible. This is editor-only, and only when the block is selected.

This regressed at some point, a while ago, probably around the introduction of extra on-click padding to show the dashed outlines of child elements.

This PR shuffles the rules a bit, reduces some of their specificity, and applies the left and right padding elsewhere to make it work.