Update `LinkControl` component to utilitse dynamic settings for additional settings "drawer" (#18285)
* Adds basic coverage test for new tab setting

* Refactor ‘new-tab` to camelCase for consistency

* Fix bug with null render if settings has a length and convert to array

Make settings an array so that it can be ordered. Fix incorrect conditional testing for .length to ensure it passes if settings _does_ have a length.

* Updates to loop over supplied settings

Previously only a “new tab” setting was hardcoded. Update so retain “new tab” as the default, but also allow for custom settings if/when provided.

* Updates docs

* Updates Nav Item Block to utilise new setting format

* Updates spacing between individual setting controls

* Update docs for settings

* Updates setting key/id to match attribute to simply and improve consistency

* Fix Nav Menu Block dynamic setting of attribute

Addresses https://github.com/WordPress/gutenberg/pull/18285#discussion_r343152809