Storybook: Components: Apply style changes (#18676)
* Storybook: Components: Apply style changes

This update fixes the Storybook x WP/Components workflow to enable for style changes made in `.scss` files to be reflected within Storybook. Previously, this was not supported because only the pre-built `.css` file was imported.

There is an interesting gotcha in this setup. The style built step leverages PostCSS for certain enhancements. For example, the `color()`. By only loading the `.scss` file, this does not work, resulting in certain styles not rendering correctly. As a work-around, I've kept one of the built `.css` files, but also loading the `.scss` file to enable the improved UI dev experience.

* Add PostCSS support. Remove importing .css as fallback

* Import block-library, block-editor, and format-library SCSS into Playground