Social Link: group all variations under one block type (#19887)
* Social Link: group all variations under one block type

* Social Link: Update test fixtures

* Address review feedback

* Fix misnamed attribute 'type' -> 'site'

* Social Link: Rename attribute 'site' to 'service'

* Social Link: Update description

* Change attribute name to service (previously site)

* Lint: Equal sign alignment

* Social Link: Remove unnecessary block deprecation

This PR introduces `core/social-link` as a new block type. Strictly
speaking, any concerns of backwards compatibility should only deal with
the migration of `core/social-link-FOO` types to this new one --
something that the parser is handling directly, populating the new
`service` attribute in that process.

Thus, no specific provisions are needed for the extant but obsolete
`site` attribute: the attribute will be silently dropped during block
validation. This can be attested by loading any content with links in
the old `core/social-link-FOO` shape.

Co-authored-by: Marcus Kazmierczak <marcus@mkaz.com>