Edit-site Package: Make public, set version to 1.3.0-alpha.0 (#20210)
This hasn't been published to npm yet. We're interested in using it to explore Full Site Editing outside of a 'regular' Gutenberg context, so we'd like to publish it.

Note that even though the package hasn't been published yet, the version number has been regularly bumped by the publish script, and it's now at 1.2.0. I'd like to set it to an alpha version (to make clear that it's still in its early stages and not yet ready for production). I've thus chosen 1.3.0-alpha.0 -- I hope that choice is okay.