Site Editor: add Fullscreen mode (#20691)
Move FullscreenMode component to block-editor package so that it can
be reused both for Post Editor and Site Editor. In order to not have this tied
to same value in both cases, Site Editor store is added as part of this change
so that its fullscreen mode preference can be saved separately.