Dismiss bottom sheet on return key submit (#20848)
* Dismiss link option bottom sheet when return is tapped

When the user creates or edits a link in text and a bottom sheet is shown, both the bottom sheet and the keyboard must be dismissed when the keyboard return key is tapped.

* Dismiss bottom sheet when Button Block keyboard is closed

There are two scenarios in regards to the Button Block because the link button and the settings button open different bottom sheets. The `isLinkSheetVisible` handles the first scenario and `closeGeneralSidebar` handles the second.