Try: Alternate 1px spacer fix. (#25528)
This PR takes a different approach to allowing a 1px tall spacer block.

The primary problem is that the sibling inserter above and below a 1px spacer block makes it appear hard to select the spacer.

This PR addresses that primarily by changing the cursor from being hardcoded to `text`, to not being changed at all. In practice that means when you try to click a 1px spacer, you're technically clicking the sibling inserter hover area, but that area redirects to the nearest block which in most cases means the spacer block.